solved Conflicts:
	src/main/java/de/metas/ui/web/WebRestApiApplication.java

+++ src/main/java/de/metas/ui/web/attachments/DocumentAttachmentsRestController.java
@@ -20,7 +20,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.google.common.collect.ImmutableList;
+
 import de.metas.attachments.AttachmentEntry;
+import de.metas.security.IUserRolePermissions;
 import de.metas.ui.web.attachments.json.JSONAttachURLRequest;
 import de.metas.ui.web.attachments.json.JSONAttachment;
 import de.metas.ui.web.exceptions.EntityNotFoundException;
@@ -71,6 +74,17 @@ public class DocumentAttachmentsRestController
 	private DocumentAttachments getDocumentAttachments(final String windowIdStr, final String documentId)
 	{
 		final DocumentPath documentPath = DocumentPath.rootDocumentPath(WindowId.fromJson(windowIdStr), documentId);
+		return getDocumentAttachments(documentPath);
+	}
+
+	private DocumentAttachments getDocumentAttachments(final DocumentPath documentPath)
+	{
+		if (documentPath.isComposedKey())
+		{
+			throw new AdempiereException("Document does not support attachments")
+					.setParameter("technicalReason", "documents with composed keys are not handled");
+		}
+
 		final TableRecordReference recordRef = documentDescriptorFactory.getTableRecordReference(documentPath);
 
 		return DocumentAttachments.builder()
@@ -117,8 +131,27 @@ public class DocumentAttachmentsRestController
 	{
 		userSession.assertLoggedIn();
 
-		return getDocumentAttachments(windowIdStr, documentId)
+		final DocumentPath documentPath = DocumentPath.rootDocumentPath(WindowId.fromJson(windowIdStr), documentId);
+		if (documentPath.isComposedKey())
+		{
+			// document with composed keys does not support attachments
+			return ImmutableList.of();
+		}
+
+		final boolean allowDelete = isAllowDeletingAttachments();
+
+		final List<JSONAttachment> attachments = getDocumentAttachments(documentPath)
 				.toJson();
+		attachments.forEach(attachment -> attachment.setAllowDelete(allowDelete));
+
+		return attachments;
+	}
+
+	private boolean isAllowDeletingAttachments()
+	{
+		return userSession
+				.getUserRolePermissions()
+				.hasPermission(IUserRolePermissions.PERMISSION_IsAttachmentDeletionAllowed);
 	}
 
 	@GetMapping("/{id}")
@@ -189,6 +222,10 @@ public class DocumentAttachmentsRestController
 	{
 		userSession.assertLoggedIn();
 
+		if (!isAllowDeletingAttachments())
+		{
+			throw new AdempiereException("Delete not allowed");
+		}
 		final DocumentId entryId = DocumentId.of(entryIdStr);
 		getDocumentAttachments(windowIdStr, documentId)
 				.deleteEntry(entryId);
